Try to make use of environment variables (plus, renaming of config/db.js)
