+++ client/app/config/index.js
@@ -1,3 +1,5 @@
-module.exports.host = 'http://localhost:5003';
-module.exports.path = '/products/';
-module.exports.apiPath = '/api/pdc/';
+module.exports.PROTOCOL = process.env.PDC_PROTOCOL || 'http';
+module.exports.HOST = process.env.PDC_HOST || 'localhost';
+module.exports.PORT = process.env.PDC_PORT || '5003';
+module.exports.PATH = '/products/';
+module.exports.API_PATH = '/api/pdc/';
