+++ server/config/server.js
@@ -1,1 +1,3 @@
-module.exports.PORT = 5003;
+module.exports.PROTOCOL = process.env.PDC_PROTOCOL || 'http';
+module.exports.HOST = process.env.PDC_HOST || 'localhost';
+module.exports.PORT = process.env.PDC_PORT || '5003';
